requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
packaging==25.0
pip-review==1.3.0
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
